theme picker refactoring
