+++ src/components/ThemePicker/ThemePicker.module.scss
@@ -1,4 +1,4 @@
-.Select {
+.ThemePicker {
   font-size: 1.3rem;
   text-align: center;
