+++ src/components/ThemePicker/utils/utils.js
@@ -1,0 +1,7 @@
+export const colors = [
+    {name: "beige", color: "beige"},
+    {name: "green", color: "#00FF7F"},
+    {name: "red", color: "#CD5C5CFF"},
+    {name: "silver", color: "#C0C0C0"},
+    {name: "blue", color: "#4169E1"},
+];
